1.01
semi-functional twitter search. currently a stream is opened and catches
all tweets that contain the query. The problem with this is it only
catches the tweets none of the background info about the tweet is
caught. This function would be more of use in a seperate app that builds
a database of all tweets that we can query.